chore: reduce bundle size use for js files (#7)
This commit changes how pure JS files are stored in the bundle. They are
no longer transpiled. Instead the `transpiled` field is now an `Option`,
and for JS files this `Option` is `None`. After the change old zips will
work with the new loader, but the old loader will not work with new
zips.