feat: Allow cloning an `EszipV2`
This patch adds a `Clone` implementation for `EszipV2`. Since this
struct can transitively contain `Waker`s, which are not clonable, this
implementation will panic if some source slot is still pending
parsing.

One issue is that the modules map of `EszipV2` is behind an
`Arc<Mutex>`, and therefore directly cloning it (with
`#[derive(Clone)]`, for example) would leave the clone "connected" to
the original, and able to change through `EszipV2::add_import_map`.
Therefore, it instead uses a custom `Clone` implementation that copies
the contents of the module map to prevent this.