chore: Reduce contention when using `EszipV2` from multiple threads
The `EszipV2` struct guards the module map behind an `Arc<Mutex>`.
This is needed to support streaming sources while returning a result
before finishing parsing. However, this means every operation that
needs access to the module map has to contend for that same lock.

This contention is inevitable while the streaming parsing is going on,
but reads also contend with each other. This change avoids this read
contention by switching the `Mutex` for a `RwLock`.

However, for pending modules, the `get_module_source` and
`get_module_source_map` methods add a waker for the current async
context to the module's waker list, and this would need some write
locking. To avoid this, this change also guards the waker list behind
a `Mutex`, which also reduces the scope of that contention by making
it specific to a module.

With these changes, given a fully-resolved `EszipV2` which isn't
modified, all operations on it are now guaranteed to be lock-free.