feat: add `take_source` & `take_source_map` methods to Module (#117)
This PR introduces methods to take the source of a module. For EszipV1,
the entire module is taken from the map, and no metadata is preserved. In
EszipV1, the caller can still access the source map after the source is
assumed.

The methods will help callers to drop source data from memory after the
the module is loaded.