feat: allow JSONC to be embedded as an import map (#129)
This commit allows JSONC to be embedded in an eszip as an import map.
This is primarily for deno's config file, which can be JSONC (e.g.
`deno.jsonc`).

Embedded JSONC data can then be retrieved by calling `get_import_map`
method, which is a new method. Keep in mind that the existing similar
method, `get_module`, will not return JSONC. This separation ensures
that whatever returned from `get_module` is definitely a valid module,
which may be useful when implementing support for dynamic import.

Ideally eszip's serialization format should have a dedicated optional
section for an import map so that we can treat import maps and modules
differently. We can't do this while keeping v2 since this would break
backward compatibility.

ref https://github.com/denoland/deploy_feedback/issues/344
ref https://github.com/denoland/fresh/pull/1188